+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.3.6] - 2026-07-15
+
+### Fixed
+- Fix estimated duration badge on custom substances with duration data
+
+## [0.3.5] - 2026-07-14
+
+### Fixed
+- No more flicker when browsing substance pages from Library
+
 ## [0.3.4] - 2026-07-14
 
 ### Added